update model functions for different queries
