+++ server/app.js
@@ -1,6 +1,6 @@
-const mysql = require("mysql2");
 const express = require("express");
 const cors = require("cors");
+const connection = require("../server/models/db.js");
 
 const app = express();
 app.use(cors());
@@ -13,13 +13,6 @@ app.get("/", (req, res) => {
 {/* API Endpoint for warehouse */}
 app.get("/warehouse", async (req, res) => {
   try {
-    const connection = mysql.createConnection({
-      host: "localhost",
-      user: "isys2099_group9_app_whadmin_user",
-      password: "jlAfD3sBdpJvz0phULIq4CAaAMIIoGNA",
-      database: "isys2099_group9_app",
-    });
-
     await new Promise((resolve, reject) => {
       connection.connect(err => {
         if (err) {
@@ -64,13 +57,6 @@ app.get("/warehouse", async (req, res) => {
 {/* API Endpoint for product */}
 app.get("/product", async (req, res) => {
   try {
-    const connection = mysql.createConnection({
-      host: "localhost",
-      user: "isys2099_group9_app_buyer_user",
-      password: "gxQJCza0eADkT5AKmeE865ZN8p1nBsar",
-      database: "isys2099_group9_app",
-    });
-
     await new Promise((resolve, reject) => {
       connection.connect(err => {
         if (err) {
