+++ server/package.json
@@ -27,6 +27,7 @@
   "homepage": "https://github.com/miketvo/rmit-isys2099-group9-app#readme",
   "dependencies": {
     "body-parser": "^1.20.2",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
